aw-server: Add Into trait for DatastoreError to HttpErrorJson
