+++ aw-server/src/endpoints/query.rs
@@ -14,19 +14,8 @@ pub fn query(
     let query_code = query_req.0.query.join("\n");
     let intervals = &query_req.0.timeperiods;
     let mut results = Vec::new();
+    let datastore = endpoints_get_lock!(state.datastore);
     for interval in intervals {
-        // Cannot re-use endpoints_get_lock!() here because it returns Err(Status) on failure and this
-        // function returns HttpResponse
-        let datastore = match state.datastore.lock() {
-            Ok(ds) => ds,
-            Err(e) => {
-                warn!("Taking datastore lock failed, returning 500: {}", e);
-                return Err(HttpErrorJson::new(
-                    Status::ServiceUnavailable,
-                    "Taking datastore lock failed, see aw-server logs".to_string(),
-                ));
-            }
-        };
         let result = match aw_query::query(&query_code, &interval, &datastore) {
             Ok(data) => data,
             Err(e) => {
